fixed navigation bar styling
