+++ src/App.css
@@ -70,16 +70,17 @@ header {
   color: #ffd600;
 }
 
+header:hover {
+  color: #ffc400; /* I didn't know which color to pick, so please help me fix this later */
+  transition: 0.3s;
+}
+
 h1 {
   font-family: "Roboto", sans-serif;
   font-weight: 500;
   font-size: 36px;
   font-style: italic;
   color: #2ba6eb;
-}
-
-h1.home {
-  margin-left: 25px;
 }
 
 h2 {
@@ -141,6 +142,13 @@ p {
 }
 
 div.home {
+  float: left;
   width: 55%;
   padding-left: 25px;
+}
+
+img.home {
+  float: right;
+  width: 45%;
+  padding-right: 25px;
 }
