--- src/components/error.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Error = () => {
-  return (
-    <div>
-      <p>Error: Path does not exist</p>
-    </div>
-  );
-};
-
-export default Error;
