+++ src/views/error.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Error = () => {
+  return (
+    <div>
+      <p>Error: Path does not exist</p>
+    </div>
+  );
+};
+
+export default Error;
